requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.2.16
langchain-community==0.2.17
langchain-google-genai==1.0.10
langgraph==0.2.55
pymongo==4.6.0
pydantic==2.4.2
pydantic-settings==2.0.3
loguru==0.7.2
google-generativeai==0.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
resend==0.6.0
python-dotenv==1.0.0
email-validator==2.1.0
